+++ packages/core/src/components/hero-image/hero-image.tsx
@@ -2,7 +2,7 @@ import { Component, ComponentInterface, Host, h } from "@stencil/core";
 
 /**
  * @slot image - Element that provides the background image, e.g. a `div` with a custom CSS class that sets `background-image`.
- * @slot - Rich content that is shown in a box on top of the image.
+ * @slot - The Highlight Box with Rich Content that is shown on top of or below the image.
  */
 @Component({
   tag: "dso-hero-image",