+++ packages/dso-toolkit/src/components/hero-image/hero-image.args.ts
@@ -1,7 +1,5 @@
 import { ArgTypes } from "storybook/internal/types";
 
-import { HeroImage } from "./hero-image.models.js";
-
 export interface HeroImageArgs {
   imageUrl: string;
 }
@@ -17,9 +15,3 @@ export const heroImageArgTypes: ArgTypes<HeroImageArgs> = {
     },
   },
 };
-
-export function heroImageArgsMapper(a: HeroImageArgs): HeroImage {
-  return {
-    ...a,
-  };
-}