+++ storybook/src/example-pages/toepassingen/aanvragen/landingspagina.stories.ts
@@ -7,10 +7,6 @@ import { mainMenu } from "../../content/main-menu.content";
 import { footerPartial } from "../../partials/footer";
 import { headerPartial } from "../../partials/header";
 
-/**
- * @deprecated Uses deprecated CSS classes (e.g., `.dso-banner`).
- * The `.dso-banner` selector will be removed in `#3923`.
- */
 const meta: Meta = {
   title: "Voorbeeldpagina's/Toepassingen/Aanvragen/Landingspagina",
 };
@@ -24,7 +20,7 @@ const Landingspagina = examplePageStories((templates) => {
     <div class="container">
       ${headerPartial(templates, { ...header, mainMenu: mainMenu("Aanvragen") })}
       <main>
-        <!-- START DEPRECATED: het <dso-hero-image> Web Component -->
+        <!-- START DEPRECATED: use <dso-hero-image> -->
         <div
           class="row dso-banner dso-banner-implementation-specific-image"
           style="background-image: url('images/hands-on-trackpad.jpg'); padding-block-end: 16px;"