+++ storybook/src/example-pages/toepassingen/loket/homepage.stories.ts
@@ -9,10 +9,6 @@ import { headerPartial } from "../../partials/header";
 
 import { linkList } from "./homepage.content";
 
-/**
- * @deprecated Uses deprecated CSS classes (e.g., `.dso-banner`).
- * The `.dso-banner` selector will be removed in `#3923`.
- */
 const meta: Meta = {
   title: "Voorbeeldpagina's/Toepassingen/Loket/Homepage",
 };
@@ -39,7 +35,7 @@ const Homepage = examplePageStories((templates) => {
       <div id="navigation"><!-- for skiplink --></div>
       ${headerPartial(templates, { ...header, mainMenu: mainMenu("Regels op de kaart") })}
       <main id="main">
-        <!-- START DEPRECATED: het <dso-hero-image> Web Component -->
+        <!-- START DEPRECATED: use <dso-hero-image> -->
         <div
           class="row dso-banner no-button-banner my-beautiful-image"
           style="background-image: url('images/hero2.jpeg')"