+++ storybook/src/example-pages/toepassingen/maatregelen-op-maat/landingspagina.stories.ts
@@ -9,10 +9,6 @@ import { headerPartial } from "../../partials/header";
 
 import { accordionSections } from "./landingspagina.content";
 
-/**
- * @deprecated Uses deprecated CSS classes (e.g., `.dso-banner`).
- * The `.dso-banner` selector will be removed in `#3923`.
- */
 const meta: Meta = {
   title: "Voorbeeldpagina's/Toepassingen/Maatregelen op maat/Landingspagina",
 };
@@ -26,7 +22,7 @@ const Landingspagina = examplePageStories((templates) => {
     <div class="container">
       ${headerPartial(templates, { ...header, mainMenu: mainMenu("Maatregelen op maat") })}
       <main>
-        <!-- START DEPRECATED: het <dso-hero-image> Web Component -->
+        <!-- START DEPRECATED: use <dso-hero-image> -->
         <div
           class="row dso-banner dso-banner-implementation-specific-image"
           style="background-image: url('images/hero4.jpeg')"